+++ Core/Components/IStrategyContext.cs
@@ -4,6 +4,6 @@ namespace Core.Components
 {
     public interface IStrategyContext
     {
-        int GetCounter(Game game, Deal deal);
+        double GetCounter(Game game, Deal deal);
     }
 }
