+++ Strategies/StrategyContexts/UstonSS/UstonSSStrategyContext.cs
@@ -10,7 +10,7 @@ namespace Strategies.StrategyContexts.UstonSS
         {
         }
 
-        protected override int GetRunningCounter(Game game) => game.History
+        protected override double GetRunningCounter(Game game) => game.History
             .Where(d => d.IsEnded)
             .Aggregate(-4 * DeckAmount, (sum, nextDeal) => sum + CardCounter.Count(nextDeal));
     }
